Yarn upgrade and refactor episodes to own file
