+++ src/data/episodes.js
@@ -1,0 +1,2 @@
+import Series from "./series";
+export default Series.map(s => s.episodes).reduce((a, b) => a.concat(b));
